cleaned KBandBandResult
