+++ wannierberri/result/__init__.py
@@ -20,5 +20,4 @@ from .__result import Result
 from .__resultdict import  ResultDict
 from .__energyresult import EnergyResult
 from .__tabresult import TABresult
-from .__kbandresult import KBandResult
-from .__kbandbandresult import KBandBandResult
+from .__kbandresult import KBandResult, KBandBandResult
