+++ wannierberri/result/__kbandresult.py
@@ -6,7 +6,7 @@ import itertools
 
 class KBandResult(Result):
 
-    def __init__(self, data, TRodd, Iodd):
+    def __init__(self, data, TRodd=None, Iodd=None):
         if isinstance(data, list):
             self.data_list = data
         else:
@@ -40,10 +40,10 @@ class KBandResult(Result):
 
     def __add__(self, other):
         assert self.fit(other)
-        return KBandResult(self.data_list + other.data_list, self.TRodd, self.Iodd)
+        return self.__class__(self.data_list + other.data_list, self.TRodd, self.Iodd)
 
     def __mul__(self, number):
-        return KBandResult([d * number for d in self.data_list], self.TRodd, self.Iodd)
+        return self.__class__([d * number for d in self.data_list], self.TRodd, self.Iodd)
 
     def __truediv__(self, number):
         return self * 1  # actually a copy
@@ -51,10 +51,10 @@ class KBandResult(Result):
     def to_grid(self, k_map):
         dataall = self.data
         data = np.array([sum(dataall[ik] for ik in km) / len(km) for km in k_map])
-        return KBandResult(data, self.TRodd, self.Iodd)
+        return self.__class__(data, self.TRodd, self.Iodd)
 
     def select_bands(self, ibands):
-        return KBandResult(self.data[:, ibands], self.TRodd, self.Iodd)
+        return self.__class__(self.data[:, ibands], self.TRodd, self.Iodd)
 
     def average_deg(self, deg):
         for i, D in enumerate(deg):
@@ -65,7 +65,7 @@ class KBandResult(Result):
 
     def transform(self, sym):
         data = [sym.transform_tensor(data, rank=self.rank, TRodd=self.TRodd, Iodd=self.Iodd) for data in self.data_list]
-        return KBandResult(data, self.TRodd, self.Iodd)
+        return self.__class__(data, self.TRodd, self.Iodd)
 
     def get_component_list(self):
         return ["".join(s) for s in itertools.product(*[("x", "y", "z")] * self.rank)]
@@ -136,3 +136,36 @@ class NoComponentError(RuntimeError):
         super().__init__("component {} does not exist for tensor with dimension {}".format(comp, dim))
 
 
+class KBandBandResult(KBandResult):
+
+    def fit(self, other):
+        for var in ['rank', 'nband','nband2']:
+            if getattr(self, var) != getattr(other, var):
+                return False
+        return True
+
+    @property
+    def rank(self):
+        return len(self.data_list[0].shape) - 3
+
+    @property
+    def nband2(self):
+        return self.data_list[0].shape[2]
+
+    def select_bands2(self, ibands2):
+        return KBandBandResult(self.data[:, :, ibands2])
+
+    def average_deg(self, deg):
+        pass   # do not average over degenerate
+        return self
+
+    def transform(self, sym):
+        return self
+
+    @property
+    def allow_sym(self):
+        return False
+
+    @property
+    def allow_frmsf(self):
+        return False
